Adds COTS probability raster prediction
Implements functionality to predict the probability of problematic COTS presence using a trained XGBoost classification model and raster predictors.

Saves predictor data to a GeoTIFF and defines a prediction function for `terra::predict`. Generates a probability raster and outputs the raster as a cloud-optimized GeoTIFF.

The new raster provides a spatial prediction of COTS problem likelihood.